test(init): centralize onboarding detector coverage
