+++ cli/src/test/kotlin/me/liam/microsmith/cli/init/OnboardingProfileDetectorTests.kt
@@ -3,8 +3,18 @@ package me.liam.microsmith.cli.init
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
+import java.io.IOException
+import java.io.UncheckedIOException
+import java.nio.file.FileSystems
+import java.nio.file.Files
 import java.nio.file.Path
+import kotlin.io.path.ExperimentalPathApi
+import kotlin.io.path.createDirectories
+import kotlin.io.path.createTempDirectory
+import kotlin.io.path.deleteRecursively
+import kotlin.io.path.writeText
 
+@OptIn(ExperimentalPathApi::class)
 class OnboardingProfileDetectorTests :
     StringSpec({
         "keeps a specific profile when multiple matchers map to the same profile" {
@@ -165,6 +175,253 @@ class OnboardingProfileDetectorTests :
 
             error.message.shouldContain("python")
         }
+
+        "ignores .NET marker traversal failures while detecting other repository markers" {
+            val nodeRoot = createTempDirectory("microsmith-init-detect-node-traversal-failure")
+            try {
+                nodeRoot.resolve("package.json").writeText("""{"name":"fixture-node"}""")
+
+                detectOnboardingProfile(
+                    projectRoot = nodeRoot,
+                    detector = detectorWithDotnetMarkerFinder { throw IOException("permission denied") },
+                ) shouldBe
+                    OnboardingProfileDetection(
+                        profile = NodeOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("package.json"),
+                    )
+            } finally {
+                runCatching { nodeRoot.deleteRecursively() }
+            }
+        }
+
+        "ignores unchecked .NET marker traversal failures while detecting other repository markers" {
+            val nodeRoot = createTempDirectory("microsmith-init-detect-node-unchecked-traversal-failure")
+            try {
+                nodeRoot.resolve("package.json").writeText("""{"name":"fixture-node"}""")
+
+                detectOnboardingProfile(
+                    projectRoot = nodeRoot,
+                    detector = detectorWithDotnetMarkerFinder {
+                        throw UncheckedIOException(IOException("permission denied"))
+                    },
+                ) shouldBe
+                    OnboardingProfileDetection(
+                        profile = NodeOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("package.json"),
+                    )
+            } finally {
+                runCatching { nodeRoot.deleteRecursively() }
+            }
+        }
+
+        ".NET detection skips unreadable directories".config(enabled = supportsPosixPermissions()) {
+            val dotnetRoot = createTempDirectory("microsmith-init-detect-dotnet-unreadable-directory")
+            val unreadableDirectory = dotnetRoot.resolve("restricted")
+            val readableProject = dotnetRoot.resolve("src/apps/service/Fixture.csproj")
+            val unreadableProject = unreadableDirectory.resolve("Ignored.csproj")
+            readableProject.parent.createDirectories()
+            unreadableDirectory.createDirectories()
+            readableProject.writeText("<Project Sdk=\"Microsoft.NET.Sdk\" />\n")
+            unreadableProject.writeText("<Project Sdk=\"Microsoft.NET.Sdk\" />\n")
+
+            val originalPermissions = Files.getPosixFilePermissions(unreadableDirectory)
+            Files.setPosixFilePermissions(unreadableDirectory, emptySet())
+            try {
+                detectOnboardingProfile(dotnetRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = DotnetOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("src/apps/service/Fixture.csproj"),
+                    )
+            } finally {
+                runCatching {
+                    Files.setPosixFilePermissions(unreadableDirectory, originalPermissions)
+                }
+                runCatching { dotnetRoot.deleteRecursively() }
+            }
+        }
+
+        listOf("pyproject.toml", "requirements.txt", "setup.py", "setup.cfg").forEach { markerFileName ->
+            "detects Python repositories from $markerFileName" {
+                val pythonRoot = createTempDirectory("microsmith-init-detect-python")
+                try {
+                    pythonRoot.resolve(markerFileName).writeText("# marker\n")
+
+                    detectOnboardingProfile(pythonRoot) shouldBe
+                        OnboardingProfileDetection(
+                            profile = PythonOnboardingProfile,
+                            selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                            matchedMarkers = listOf(markerFileName),
+                        )
+                } finally {
+                    runCatching { pythonRoot.deleteRecursively() }
+                }
+            }
+        }
+
+        "keeps the Python profile when multiple Python markers match" {
+            val pythonRoot = createTempDirectory("microsmith-init-detect-python-multi")
+            try {
+                pythonRoot.resolve("pyproject.toml").writeText("[project]\nname = \"fixture-python\"\n")
+                pythonRoot.resolve("requirements.txt").writeText("protobuf==0.0.0\n")
+
+                detectOnboardingProfile(pythonRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = PythonOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("pyproject.toml", "requirements.txt"),
+                    )
+            } finally {
+                runCatching { pythonRoot.deleteRecursively() }
+            }
+        }
+
+        "detects Rust repositories from Cargo.toml at package or workspace roots" {
+            val packageRoot = createTempDirectory("microsmith-init-detect-rust-package")
+            val workspaceRoot = createTempDirectory("microsmith-init-detect-rust-workspace")
+            try {
+                packageRoot.resolve("Cargo.toml").writeText(
+                    """
+                    [package]
+                    name = "fixture-rust"
+                    version = "0.1.0"
+                    edition = "2024"
+                    """.trimIndent() + "\n",
+                )
+                workspaceRoot.resolve("crates/app").createDirectories()
+                workspaceRoot.resolve("Cargo.toml").writeText(
+                    """
+                    [workspace]
+                    members = ["crates/app"]
+                    """.trimIndent() + "\n",
+                )
+
+                detectOnboardingProfile(packageRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = RustOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("Cargo.toml"),
+                    )
+                detectOnboardingProfile(workspaceRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = RustOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("Cargo.toml"),
+                    )
+            } finally {
+                runCatching { packageRoot.deleteRecursively() }
+                runCatching { workspaceRoot.deleteRecursively() }
+            }
+        }
+
+        "falls back to the generic profile when a nested Rust crate exists without a root Cargo.toml" {
+            val repoRoot = createTempDirectory("microsmith-init-detect-rust-nested-only")
+            try {
+                repoRoot.resolve("crates/app").createDirectories()
+                repoRoot.resolve("crates/app/Cargo.toml").writeText(
+                    """
+                    [package]
+                    name = "fixture-rust-app"
+                    version = "0.1.0"
+                    edition = "2024"
+                    """.trimIndent() + "\n",
+                )
+
+                detectOnboardingProfile(repoRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = GenericOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.NO_MARKERS_MATCHED,
+                        matchedMarkers = emptyList(),
+                    )
+            } finally {
+                runCatching { repoRoot.deleteRecursively() }
+            }
+        }
+
+        "detects built-in repository profiles and falls back to the generic profile for ambiguous markers" {
+            val nodeRoot = createTempDirectory("microsmith-init-detect-node")
+            val goRoot = createTempDirectory("microsmith-init-detect-go")
+            val pythonRoot = createTempDirectory("microsmith-init-detect-python")
+            val rubyRoot = createTempDirectory("microsmith-init-detect-ruby")
+            val rustRoot = createTempDirectory("microsmith-init-detect-rust")
+            val dotnetRoot = createTempDirectory("microsmith-init-detect-dotnet")
+            val mixedRoot = createTempDirectory("microsmith-init-detect-mixed")
+            try {
+                nodeRoot.resolve("package.json").writeText("""{"name":"fixture-node"}""")
+                goRoot.resolve("go.mod").writeText("module example.com/microsmith/fixture\n")
+                pythonRoot.resolve("pyproject.toml").writeText("[project]\nname = \"fixture-python\"\n")
+                rubyRoot.resolve("gems.rb").writeText("source \"https://rubygems.org\"\n")
+                rustRoot.resolve("Cargo.toml").writeText("[package]\nname = \"fixture-rust\"\nversion = \"0.1.0\"\n")
+                dotnetRoot.resolve("src/apps/service").createDirectories()
+                dotnetRoot.resolve("src/apps/service/Fixture.csproj")
+                    .writeText("<Project Sdk=\"Microsoft.NET.Sdk\" />\n")
+                mixedRoot.resolve("gems.rb").writeText("source \"https://rubygems.org\"\n")
+                mixedRoot.resolve("Cargo.toml").writeText("[package]\nname = \"fixture-rust\"\nversion = \"0.1.0\"\n")
+
+                detectOnboardingProfile(nodeRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = NodeOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("package.json"),
+                    )
+                detectOnboardingProfile(goRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = GoOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("go.mod"),
+                    )
+                detectOnboardingProfile(pythonRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = PythonOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("pyproject.toml"),
+                    )
+                detectOnboardingProfile(rubyRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = RubyOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("gems.rb"),
+                    )
+                detectOnboardingProfile(rustRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = RustOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("Cargo.toml"),
+                    )
+                detectOnboardingProfile(dotnetRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = DotnetOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.MATCHED_PROFILE,
+                        matchedMarkers = listOf("src/apps/service/Fixture.csproj"),
+                    )
+                detectOnboardingProfile(mixedRoot) shouldBe
+                    OnboardingProfileDetection(
+                        profile = GenericOnboardingProfile,
+                        selectionReason = OnboardingProfileSelectionReason.AMBIGUOUS_MARKERS,
+                        matchedMarkers = listOf("Cargo.toml", "gems.rb"),
+                    )
+            } finally {
+                runCatching { nodeRoot.deleteRecursively() }
+                runCatching { goRoot.deleteRecursively() }
+                runCatching { pythonRoot.deleteRecursively() }
+                runCatching { rubyRoot.deleteRecursively() }
+                runCatching { rustRoot.deleteRecursively() }
+                runCatching { dotnetRoot.deleteRecursively() }
+                runCatching { mixedRoot.deleteRecursively() }
+            }
+        }
     })
 
 private val UNUSED_PROJECT_ROOT: Path = Path.of(".")
+
+private fun supportsPosixPermissions(): Boolean =
+    !runningOnWindows() && FileSystems.getDefault().supportedFileAttributeViews().contains("posix")
+
+private fun detectorWithDotnetMarkerFinder(dotnetMarkerFinder: (Path) -> String?): OnboardingProfileDetector =
+    OnboardingProfileDetector(
+        matchers = BuiltInOnboardingProfileMatchers.all(dotnetMarkerFinder),
+    )
+
+private fun runningOnWindows(): Boolean = System.getProperty("os.name").startsWith("Windows", ignoreCase = true)
